# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: lembar API
  version: 0.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >-
    Baseline contract for planning. The backend repository owns the executable OpenAPI
    document. This file does not authorize implementation of every path at once.
servers:
  - url: /
    description: Current API origin or configured same-origin gateway
...
paths:
...
  /v1/auth/register:
    post:
      summary: Register a new user account
      tags: [Identity]
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string, minLength: 1 }
                roles:
                  type: array
                  items: { type: string }
                  default: ["subscriber"]
                  description: User roles (default ["subscriber"])
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/login:
    post:
      summary: Login and receive JWT token
      tags: [Identity]
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/auth/me:
    get:
      summary: Get current authenticated user info
      tags: [Identity]
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/me/roles:
    patch:
      summary: Update user roles (superadmin only)
      tags: [Identity]
      operationId: updateUserRoles
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [roles]
              properties:
                roles:
                  type: array
                  items: { type: string }
                  description: New roles array
      responses:
        "200":
          description: Roles updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfoResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Forbidden - superadmin role required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/me:
    get:
      summary: Read the current account and workspace options
      tags: [Identity]
      operationId: getCurrentAccount
      responses:
        "200":
          description: Current account and available workspaces
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/dashboard/summary:
    get:
      summary: Read the active workspace dashboard summary
      tags: [Identity]
      operationId: getDashboardSummary
      responses:
        "200":
          description: Active workspace dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummaryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/catalog/grades:
    get:
      summary: List available grades
      tags: [Catalog]
      operationId: listGrades
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: Available grades
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CatalogOption"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/catalog/subjects:
    get:
      summary: List subjects for a grade and curriculum version
      tags: [Catalog]
      operationId: listSubjects
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - in: query
          name: gradeId
          required: true
          schema: { type: string }
        - in: query
          name: curriculumVersionId
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Available subjects
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CatalogOption"
        "400":
          $ref: "#/components/responses/ValidationError"
  /v1/catalog/materials:
    get:
      summary: List materials for the selected catalog context
      tags: [Catalog]
      operationId: listMaterials
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - in: query
          name: gradeId
          required: true
          schema: { type: string }
        - in: query
          name: subjectId
          required: true
          schema: { type: string }
        - in: query
          name: curriculumVersionId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Available materials
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CatalogOption"
        "400":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-lembar_session
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token dengan format "Bearer <token>"
  parameters:
    WorkspaceId:
      in: header
      name: X-Workspace-Id
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    ValidationError:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId, retryable]
          properties:
            code:
              type: string
              enum:
                - AUTH_REQUIRED
                - PERMISSION_DENIED
                - VALIDATION_FAILED
                - RESOURCE_NOT_FOUND
                - STATE_CONFLICT
                - INTERNAL_ERROR
                - WORKSPACE_ACCESS_DENIED
                - RATE_LIMITED
            message: { type: string }
            fieldErrors:
              type: object
              additionalProperties:
                type: array
                items: { type: string }
            requestId: { type: string }
            retryable: { type: boolean }
    AuthTokenResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT token for authentication
        user:
          type: object
          required: [id, email, name, roles]
          properties:
            id: { type: string, format: uuid }
            email: { type: string, format: email }
            name: { type: string }
            roles:
              type: array
              items: { type: string }
              description: User roles (e.g., ["subscriber"], ["teacher", "school_admin"])
            workspaceId: { type: string, format: uuid, nullable: true }
    UserInfoResponse:
      type: object
      required: [id, email, name, roles, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        roles:
          type: array
          items: { type: string }
          description: User roles array
        workspaceId: { type: string, format: uuid, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    WorkspaceSummary:
      type: object
      required: [id, type, name, role]
      properties:
        id: { type: string }
        type: { type: string, enum: [personal, school] }
        name: { type: string }
        role: { type: string, enum: [teacher, school_admin, subscriber, superadmin] }
        permissions:
          type: array
          items:
            type: string
            enum:
              - assessment.create
              - assessment.read
              - assessment.review
              - assessment.finalize
              - source.manage
              - library.manage
              - workspace.member.manage
              - workspace.usage.read
              - platform.support.act
        isActive: { type: boolean }
    MeResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [account, workspaces]
          properties:
            account:
              type: object
              required: [id, displayName]
              properties:
                id: { type: string }
                displayName: { type: string }
            activeWorkspaceId: { type: string }
            activeWorkspace: { $ref: "#/components/schemas/WorkspaceSummary" }
            context:
              type: object
              required: [workspaceIds, permissionSet]
              properties:
                workspaceIds:
                  type: array
                  items: { type: string }
                permissionSet:
                  type: array
                  items:
                    type: string
            workspaces:
              type: array
              items: { $ref: "#/components/schemas/WorkspaceSummary" }
    DashboardSummaryResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [workspace, metrics, emptyState]
          properties:
            workspace: { $ref: "#/components/schemas/WorkspaceSummary" }
            metrics:
              type: object
              required: [assessments, sources, jobs]
              properties:
                assessments:
                  type: object
                  required: [total, draft, inReview, final]
                  properties:
                    total: { type: integer, minimum: 0 }
                    draft: { type: integer, minimum: 0 }
                    inReview: { type: integer, minimum: 0 }
                    final: { type: integer, minimum: 0 }
                sources:
                  type: object
                  required: [total, ready, processing, failed]
                  properties:
                    total: { type: integer, minimum: 0 }
                    ready: { type: integer, minimum: 0 }
                    processing: { type: integer, minimum: 0 }
                    failed: { type: integer, minimum: 0 }
                jobs:
                  type: object
                  required: [total, active, failed]
                  properties:
                    total: { type: integer, minimum: 0 }
                    active: { type: integer, minimum: 0 }
                    failed: { type: integer, minimum: 0 }
            emptyState:
              type: object
              required: [isEmpty, message]
              properties:
                isEmpty: { type: boolean }
                message: { type: string }
    CatalogOption:
      type: object
      required: [id, label, status]
      properties:
        id: { type: string }
        label: { type: string }
        status: { type: string, enum: [active, archived, unavailable] }
...
security:
  - BearerAuth: []